refactor: update settings page
