+++ .eslintrc.js
@@ -1,29 +1,29 @@
 module.exports = {
-  root: true,
-  parser: '@typescript-eslint/parser',
-  plugins: ['react', '@typescript-eslint', 'prettier', 'react-hooks'],
-  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
-  rules: {
-    '@typescript-eslint/no-unused-vars': [2, { args: 'all', argsIgnorePattern: '^_' }],
-    '@typescript-eslint/no-empty-interface': ['off'],
-    '@typescript-eslint/no-explicit-any': ['off'],
-    'react/react-in-jsx-scope': 'off',
-    '@typescript-eslint/no-namespace': 'off',
-    'prettier/prettier': 'error',
-    'react-hooks/rules-of-hooks': 'error',
-    'react-hooks/exhaustive-deps': 'warn',
-  },
-  env: {
-    browser: true,
-    commonjs: true,
-    es2021: true,
-    node: true,
-  },
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
+    root: true,
+    parser: '@typescript-eslint/parser',
+    plugins: ['react', '@typescript-eslint', 'prettier', 'react-hooks'],
+    extends: ['eslint:recommended'],
+    rules: {
+        '@typescript-eslint/no-unused-vars': [2, { args: 'all', argsIgnorePattern: '^_' }],
+        '@typescript-eslint/no-empty-interface': ['off'],
+        '@typescript-eslint/no-explicit-any': ['off'],
+        'react/react-in-jsx-scope': 'off',
+        '@typescript-eslint/no-namespace': 'off',
+        'prettier/prettier': 'off',
+        'react-hooks/rules-of-hooks': 'error',
+        'react-hooks/exhaustive-deps': 'warn',
     },
-    ecmaVersion: 'latest',
-    sourceType: 'module',
-  },
+    env: {
+        browser: true,
+        commonjs: true,
+        es2021: true,
+        node: true,
+    },
+    parserOptions: {
+        ecmaFeatures: {
+            jsx: true,
+        },
+        ecmaVersion: 'latest',
+        sourceType: 'module',
+    },
 };
