+++ .prettierrc.js
@@ -1,11 +1,12 @@
 module.exports = {
-  printWidth: 120,
-  tabWidth: 2,
-  semi: true,
-  singleQuote: true,
-  quoteProps: 'as-needed',
-  trailingComma: 'all',
-  bracketSpacing: true,
-  arrowParens: 'always',
-  endOfLine: 'auto',
+    printWidth: 800,
+    useTabs: false,
+    tabWidth: 4,
+    semi: true,
+    singleQuote: true,
+    quoteProps: 'as-needed',
+    trailingComma: 'all',
+    bracketSpacing: true,
+    arrowParens: 'always',
+    endOfLine: 'auto',
 };
