--- src/constants.ts
@@ -1,2 +1,0 @@
-export const MEMOS_VIEW_TYPE = 'memos_view';
-export const VIEW_TYPE = 'react-view';
