--- src/desktop.ini
@@ -1,4 +1,0 @@
-[ViewState]
-Mode=
-Vid=
-FolderType=Generic
