+++ src/obComponents/obPinQuery.ts
@@ -1,6 +1,6 @@
 import { moment, Notice, TFile } from 'obsidian';
 import appStore from '../stores/appStore';
-import { QueryFileName } from '../memos';
+import { QueryFileName } from '../memosView';
 import { getDailyNotePath } from '../helpers/utils';
 
 export const pinQueryInFile = async (queryID: string): Promise<any> => {
