+++ src/pages/Setting.tsx
@@ -11,37 +11,37 @@ import MenuSvg from '../icons/menu.svg?component';
 interface Props {}
 
 const Setting: React.FC<Props> = () => {
-  const {
-    globalState: { isMobileView },
-  } = useContext(appContext);
+    const {
+        globalState: { isMobileView },
+    } = useContext(appContext);
 
-  useEffect(() => {
-    memoService.fetchAllMemos();
-  }, []);
+    useEffect(() => {
+        memoService.fetchAllMemos();
+    }, []);
 
-  const handleShowSidebarBtnClick = useCallback(() => {
-    globalStateService.setShowSiderbarInMobileView(true);
-  }, []);
+    const handleShowSidebarBtnClick = useCallback(() => {
+        globalStateService.setShowSidebarInMobileView(true);
+    }, []);
 
-  return (
-    <div className="preference-wrapper">
-      <div className="section-header-container">
-        <div className="title-text">
-          <Only when={isMobileView}>
-            <button className="action-btn" onClick={handleShowSidebarBtnClick}>
-              {/*<img className="icon-img" src={menuSvg} alt="menu" />*/}
-              <MenuSvg className="icon-img" />
-            </button>
-          </Only>
-          <span className="normal-text">账号与设置</span>
+    return (
+        <div className="preference-wrapper">
+            <div className="section-header-container">
+                <div className="title-text">
+                    <Only when={isMobileView}>
+                        <button className="action-btn" onClick={handleShowSidebarBtnClick}>
+                            {/*<img className="icon-img" src={menuSvg} alt="menu" />*/}
+                            <MenuSvg className="icon-img" />
+                        </button>
+                    </Only>
+                    <span className="normal-text">账号与设置</span>
+                </div>
+            </div>
+
+            <div className="sections-wrapper">
+                <PreferencesSection />
+            </div>
         </div>
-      </div>
-
-      <div className="sections-wrapper">
-        <PreferencesSection />
-      </div>
-    </div>
-  );
+    );
 };
 
 export default Setting;
