+++ src/translations/locale/pt-br.ts
@@ -2,7 +2,7 @@
 // Brazilian Portuguese
 
 export default {
-  // setting.ts
+  // memosSetting.ts
   welcome: 'Bem-vindo ao Memos!',
   ribbonIconTitle: 'Memos',
   months: [
